+++ server/core/controllers/api/videos/embed-privacy.ts
@@ -76,9 +76,15 @@ async function isVideoEmbedOnDomainAllowed (req: express.Request, res: express.R
   const video = res.locals.videoWithBlacklist
   const isOnInstance = req.query.domain === WEBSERVER.HOST
 
-  const domainAllowed = video.embedPrivacyPolicy === VideoEmbedPrivacyPolicy.ALL_ALLOWED || isOnInstance
-    ? true
-    : await VideoEmbedPrivacyDomainModel.isDomainAllowed(video.id, req.query.domain)
+  let domainAllowed: boolean
+
+  if (video.embedPrivacyPolicy === VideoEmbedPrivacyPolicy.DISABLED) {
+    domainAllowed = false
+  } else if (video.embedPrivacyPolicy === VideoEmbedPrivacyPolicy.ALL_ALLOWED || isOnInstance) {
+    domainAllowed = true
+  } else {
+    domainAllowed = await VideoEmbedPrivacyDomainModel.isDomainAllowed(video.id, req.query.domain)
+  }
 
   const user = getAuthUser(res)
 
@@ -86,7 +92,7 @@ async function isVideoEmbedOnDomainAllowed (req: express.Request, res: express.R
     ? null
     : false
 
-  if (domainAllowed === false && user) {
+  if (domainAllowed === false && user && video.embedPrivacyPolicy !== VideoEmbedPrivacyPolicy.DISABLED) {
     userBypassAllowed = await checkCanManageVideo({
       user,
       video: await VideoModel.loadWithRights(video.id),