+++ server/core/helpers/custom-validators/video-channel-syncs.ts
@@ -1,6 +1,11 @@
-import { STREAM_SYNC_STATE } from '@server/initializers/constants.js'
+import { VideoPrivacy } from '@peertube/peertube-models'
+import { STREAM_SYNC_STATE, VIDEO_PRIVACIES } from '@server/initializers/constants.js'
 import { exists } from './misc.js'
 
 export function isVideoChannelSyncStateValid (value: any) {
   return exists(value) && STREAM_SYNC_STATE[value] !== undefined
+}
+
+export function isVideoChannelSyncPrivacyValid (value: any) {
+  return VIDEO_PRIVACIES[value] !== undefined && value !== VideoPrivacy.PASSWORD_PROTECTED
 }