+++ server/core/initializers/migrations/1080-video-channel-sync-privacy.ts
@@ -1,0 +1,22 @@
+import * as Sequelize from 'sequelize'
+
+async function up (utils: {
+  transaction: Sequelize.Transaction
+  queryInterface: Sequelize.QueryInterface
+  sequelize: Sequelize.Sequelize
+}): Promise<void> {
+  await utils.queryInterface.addColumn('videoChannelSync', 'videoPrivacy', {
+    type: Sequelize.INTEGER,
+    allowNull: true,
+    defaultValue: null
+  }, { transaction: utils.transaction })
+}
+
+function down (options) {
+  throw new Error('Not implemented.')
+}
+
+export {
+  down,
+  up
+}